requirements.txt
attrs==19.1.0
json-log-plots==0.0.1
fire==0.1.3
matplotlib==3.0.3
numpy==1.16.2
sentencepiece==0.1.8
torch>=1.0.1
tqdm==4.31.1
